+++ docs/conf.py
@@ -213,7 +213,10 @@ html_theme = "sphinx_rtd_theme"
 # theme further.  For a list of options available for each theme, see the
 # documentation.
 #
-html_theme_options = {"style_external_links": True}
+html_theme_options = {
+    "style_external_links": True,
+    "logo_only": True
+}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -221,6 +224,8 @@ html_theme_options = {"style_external_links": True}
 if not os.path.exists("_static"):
     os.makedirs("_static")
 html_static_path = ["_static"]
+
+html_logo = "_static/_images/logos/ravenpy-logo.png"
 
 # -- Options for HTMLHelp output ---------------------------------------