+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ v0.17.0 (unreleased)
 * Updated the cookiecutter template to the latest commit and synchronized dependencies between PyPI and Anaconda recipes. (PR #427)
 * Updated `ts_fit_graph` logic for `matplotlib` >= 3.10.0 compatibility. (PR #434)
 * Temporarily pinned `pygments` below v2.19 due to a breaking change affecting `sphinx-codeautolink`. (PR #434)
+* Adopted a new RavenPy logo for the documentation. (PR #428)
 
 v0.16.1 (2024-12-05)
 --------------------